refactor: DTO Article 도메인 이동
