+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/request/PostArticleRequestDto.java
@@ -1,9 +1,12 @@
-package com.icehufs.icebreaker.dto.request.article;
+package com.icehufs.icebreaker.domain.article.dto.request;
 
-import com.icehufs.icebreaker.entity.ArticleCategoryEnum;
-
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+
+import com.icehufs.icebreaker.domain.article.domain.entity.ArticleCategoryEnum;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,7 +14,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class PatchArticleRequestDto {
+public class PostArticleRequestDto {
 
     @NotBlank
     private String articleTitle;
@@ -20,5 +23,7 @@ public class PatchArticleRequestDto {
     private String articleContent;
 
     @NotNull
+    @Enumerated(EnumType.ORDINAL)
     private ArticleCategoryEnum category;
+
 }
