+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/response/CheckOwnOfArticleResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.article;
+package com.icehufs.icebreaker.domain.article.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,27 +7,28 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class DeleteArticleResponseDto extends ResponseDto {
-    private DeleteArticleResponseDto(){
+public class CheckOwnOfArticleResponseDto extends ResponseDto{
+        private CheckOwnOfArticleResponseDto() {
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
-
-        public static ResponseEntity<DeleteArticleResponseDto> success(){
-        DeleteArticleResponseDto result = new DeleteArticleResponseDto();
+    
+    public static ResponseEntity<CheckOwnOfArticleResponseDto> success() {
+        CheckOwnOfArticleResponseDto result = new CheckOwnOfArticleResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
-        public static ResponseEntity<ResponseDto> notExistUser (){
+
+    public static ResponseEntity<ResponseDto> notExistUser (){
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_USER, ResponseMessage.NOT_EXISTED_USER);
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
 
     public static ResponseEntity<ResponseDto> noExistArticle(){
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); 
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result);
     }
-    
+
     public static ResponseEntity<ResponseDto> noPermission(){
         ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSION, ResponseMessage.NO_PERMISSION);
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(result); 
-    }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(result); }
+    
 }
