+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/response/PatchArticleResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.article;
+package com.icehufs.icebreaker.domain.article.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,16 +7,14 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-import lombok.Getter;
+public class PatchArticleResponseDto extends ResponseDto{
 
-@Getter
-public class PostCommentResponseDto extends ResponseDto{
-    private PostCommentResponseDto(){
+        private PatchArticleResponseDto(){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
 
-    public static ResponseEntity<PostCommentResponseDto> success(){
-        PostCommentResponseDto result = new PostCommentResponseDto();
+        public static ResponseEntity<PatchArticleResponseDto> success(){
+            PatchArticleResponseDto result = new PatchArticleResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
         public static ResponseEntity<ResponseDto> notExistUser (){
@@ -28,5 +26,9 @@ public class PostCommentResponseDto extends ResponseDto{
         ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); }
 
-
+        
+    public static ResponseEntity<ResponseDto> noPermission(){
+        ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSION, ResponseMessage.NO_PERMISSION);
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(result); }
+    
 }
