+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/response/PatchCommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.article;
+package com.icehufs.icebreaker.domain.article.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,14 +7,14 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class PatchArticleResponseDto extends ResponseDto{
-
-        private PatchArticleResponseDto(){
+public class PatchCommentResponseDto extends ResponseDto{
+    
+        private PatchCommentResponseDto(){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
 
-        public static ResponseEntity<PatchArticleResponseDto> success(){
-            PatchArticleResponseDto result = new PatchArticleResponseDto();
+        public static ResponseEntity<PatchCommentResponseDto> success(){
+            PatchCommentResponseDto result = new PatchCommentResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
         public static ResponseEntity<ResponseDto> notExistUser (){
@@ -22,13 +22,8 @@ public class PatchArticleResponseDto extends ResponseDto{
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
 
-    public static ResponseEntity<ResponseDto> noExistArticle(){
-        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); }
-
         
     public static ResponseEntity<ResponseDto> noPermission(){
         ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSION, ResponseMessage.NO_PERMISSION);
         return ResponseEntity.status(HttpStatus.FORBIDDEN).body(result); }
-    
 }
