+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/dto/response/PostCommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.response.article;
+package com.icehufs.icebreaker.domain.article.dto.response;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,14 +7,16 @@ import com.icehufs.icebreaker.common.ResponseCode;
 import com.icehufs.icebreaker.common.ResponseMessage;
 import com.icehufs.icebreaker.dto.response.ResponseDto;
 
-public class PatchCommentResponseDto extends ResponseDto{
-    
-        private PatchCommentResponseDto(){
+import lombok.Getter;
+
+@Getter
+public class PostCommentResponseDto extends ResponseDto{
+    private PostCommentResponseDto(){
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
     }
 
-        public static ResponseEntity<PatchCommentResponseDto> success(){
-            PatchCommentResponseDto result = new PatchCommentResponseDto();
+    public static ResponseEntity<PostCommentResponseDto> success(){
+        PostCommentResponseDto result = new PostCommentResponseDto();
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
         public static ResponseEntity<ResponseDto> notExistUser (){
@@ -22,8 +24,9 @@ public class PatchCommentResponseDto extends ResponseDto{
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(result);
     }
 
-        
-    public static ResponseEntity<ResponseDto> noPermission(){
-        ResponseDto result = new ResponseDto(ResponseCode.NO_PERMISSION, ResponseMessage.NO_PERMISSION);
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(result); }
+    public static ResponseEntity<ResponseDto> noExistArticle(){
+        ResponseDto result = new ResponseDto(ResponseCode.NOT_EXISTED_ARTICLE, ResponseMessage.NOT_EXISTED_ARTICLE);
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(result); }
+
+
 }
